refactor(reasoning): WS-3 Phase 2 — move tool-gating Act→Decide (RC-2)
Per ADR adjudication 2026-05-28 option (a): tool gating is a Decide
capability concern (pre-Act selection), not an Act concern. Mechanical
move + 9 import-path updates. No logic changes.

  - Source: packages/reasoning/src/kernel/capabilities/act/tool-gating.ts
  - Target: packages/reasoning/src/kernel/capabilities/decide/tool-gating.ts
  - 9 consumer import-path updates (act.ts, guard.ts, think.ts, think-guards.ts,
    context-manager.ts, prompt-sections-default.ts, kernel-state.ts, index.ts,
    tool-utils.test.ts)
  - Architectural docstring added at top of moved file
  - git mv used to preserve history

Cycle 2 reduction: reason→act/tool-gating edges (3 files) eliminated;
reason→act/tool-execution + reason→act/tool-parsing remain (Phase 3 target).

Refs:
  - thin spec:  wiki/Planning/Implementation-Plans/2026-05-28-ws-3-kernel-capability-dag.md §Phase 2
  - ADR:        wiki/Decisions/2026-05-28-tool-gating-disposition.md (option a)
  - master:     wiki/Planning/Implementation-Plans/2026-05-28-canonical-refactor.md §4 RC-2 + §3.6 F2
  - model:      wiki/Architecture/Design-Specs/2026-05-28-canonical-architecture-model.md §2.3 + §6

Verification gates:
  - Build:      38/38 successful (turbo)
  - Tests:      5750 pass / 0 fail / 23 skip (5773 across 659 files)
  - Typecheck:  baseline preserved — only pre-existing testing#typecheck cf-24
                (Lever 8) failure remains; no new errors
  - Tool-gating targeted: 46/46 pass in tool-utils.test.ts
  - Residual:   zero "act/tool-gating" import sites remain (only the moved
                file's own historical docstring reference)
