+++ packages/reasoning/src/index.ts
@@ -161,7 +161,7 @@ export type { KernelMetaToolsConfig } from "./types/kernel-meta-tools.js";
 
 // ─── Shared Utilities ───
 export { filterToolsByRelevance } from "./kernel/capabilities/attend/tool-formatting.js";
-export { planNextMoveBatches } from "./kernel/capabilities/act/tool-gating.js";
+export { planNextMoveBatches } from "./kernel/capabilities/decide/tool-gating.js";
 export type { ToolSchema, ToolParamSchema } from "./kernel/capabilities/attend/tool-formatting.js";
 export type { KernelMessage, EntropyScoreLike } from "./kernel/state/kernel-state.js";
 export { META_TOOLS, INTROSPECTION_META_TOOLS } from "./kernel/state/kernel-constants.js";
