+++ packages/reasoning/tests/strategies/kernel/utils/tool-utils.test.ts
@@ -15,7 +15,7 @@ import {
   gateNativeToolCallsForRequiredTools,
   buildToolElaborationInjection,
   planNextMoveBatches,
-} from "../../../../src/kernel/capabilities/act/tool-gating.js";
+} from "../../../../src/kernel/capabilities/decide/tool-gating.js";
 
 describe("hasFinalAnswer", () => {
   it("returns true for FINAL ANSWER: prefix", () => {
